database: Remove global math.h from SoLazyElement.h

+++ libInventor/database/src/nodes/SoCamera.cpp
@@ -71,7 +71,8 @@
 #include <Inventor/elements/SoFocalDistanceElement.h>
 #include <Inventor/elements/SoStereoElement.h>
 #include <Inventor/nodes/SoCamera.h>
-#include <stdlib.h>
+
+#include <cmath>
 #include <algorithm>
 
 #ifdef SB_OS_WIN
@@ -191,8 +192,8 @@ SoCamera::pointAt(const SbVec3f &targetPoint)
 {
     SbVec3f	d(targetPoint - position.getValue());
 
-    SbRotation	yRot(SbVec3f(0,1,0), atan2(-d[0], -d[2]));
-    SbRotation  xRot(SbVec3f(1,0,0), atan2(d[1], sqrtf(d[0]*d[0] + d[2]*d[2])));
+    SbRotation	yRot(SbVec3f(0,1,0), std::atan2(-d[0], -d[2]));
+    SbRotation  xRot(SbVec3f(1,0,0), std::atan2(d[1], std::sqrt(d[0]*d[0] + d[2]*d[2])));
 
     orientation.setValue(xRot * yRot);
 }
